pyproject.toml
[project]
name = "aigc-kit"
version = "0.1.0"
description = "AIGC toolkit - unified image generation, LLM chat & storage"
requires-python = ">=3.12"
license = "MIT"
readme = "README.md"
authors = [{ name = "seeease" }]
keywords = ["aigc", "llm", "image-generation", "embedding", "storage"]
classifiers = [
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "httpx>=0.28",
    "google-genai>=1.63.0",
    "boto3>=1.42.49",
    "openai>=2.21.0",
    "pillow>=12.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aigc_kit"]

[tool.ruff]
target-version = "py312"

[dependency-groups]
dev = [
    "ruff>=0.15.0",
]
